requirements.txt
opencv-python
pytesseract
Pillow
numpy
easyocr
torch
torchvision
fastapi
uvicorn
pydantic
